requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/dev.in
#
absl-py==1.2.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
cycler==0.11.0
    # via matplotlib
distlib==0.3.5
    # via virtualenv
filelock==3.8.0
    # via virtualenv
flatbuffers==1.12
    # via tensorflow
fonttools==4.37.0
    # via matplotlib
gast==0.4.0
    # via tensorflow
google-auth==2.11.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.47.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.7.0
    # via tensorflow
identify==2.5.3
    # via pre-commit
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via markdown
joblib==1.1.0
    # via scikit-learn
keras==2.9.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.4.4
    # via matplotlib
libclang==14.0.6
    # via tensorflow
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via werkzeug
matplotlib==3.5.3
    # via -r requirements/base.in
nodeenv==1.7.0
    # via pre-commit
numpy==1.23.2
    # via
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.2.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   matplotlib
    #   tensorflow
pandas==1.4.3
    # via -r requirements/base.in
pillow==9.2.0
    # via matplotlib
platformdirs==2.5.2
    # via virtualenv
pre-commit==2.16.0
    # via -r requirements/dev.in
protobuf==3.19.4
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.2.1
    # via pandas
pyyaml==6.0
    # via pre-commit
requests==2.28.1
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scikit-learn==1.1.2
    # via -r requirements/base.in
scipy==1.9.1
    # via scikit-learn
six==1.16.0
    # via
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   python-dateutil
    #   tensorflow
tensorboard==2.9.1
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.9.1
    # via -r requirements/base.in
tensorflow-estimator==2.9.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.26.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
threadpoolctl==3.1.0
    # via scikit-learn
toml==0.10.2
    # via pre-commit
typing-extensions==4.3.0
    # via tensorflow
urllib3==1.26.12
    # via requests
virtualenv==20.16.3
    # via pre-commit
werkzeug==2.2.2
    # via tensorboard
wheel==0.37.1
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via tensorflow
zipp==3.8.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
